Fix template category titless
When listing proposals an s is added to the end of the category to
pluralise it. This looks rather silly if the category title already ends
in an s.

So now adding the s is only done if there isn't already one there.

This patches over the issue somewhat, proper pluralisation is hard.
Rails trys rather hard to do it and there are always bugs, the english
language is just too fucked up to do it properly.

We kind of get away with it as the categories don't change often and the
general public don't see it.