Convert h.line_break to be markupsafe
This allows h.line_break to be used as a filter, as well as as a
function.

The main advantage of this is that it allows the standard html escaping
filter to be applied, rather than requiring the naked flag (|n) to be
used. This improves #432.

I changed all the simple mako line_break invokations to use the filtered
style, I have not verified all these by hand.

In practice this makes things easier and fixes a number of issues:

* Some code manually escaped the html before using line_break and the
  naked filter. This was made considerably easier to read.

* Some code used line_break but did not enabled the naked filter. This
  would have caused the added <br> to be escaped and show up literally
  in the output... this is now fixed.

* Some code used line_break with the naked filter but did not manually
  html escape the string first, causing a potential injection attack
  site... this is now fixed.