Investigated next proposal and tweaked it
Issue #8 requests that the next random proposal be of the same type
as the current proposal. I confirmed that this is already the case.

Wrote a test for zk.model.proposal.find_next_proposal() locking in the
current behaviour.

Improved the SQL query in find_next_proposal(), removed some elements
which were used in the past but are no longer required - roughly doubles
the speed of the query.

Added another filter to the SQL query so that next proposal can't be
authored by the reviewer. Costs roughly half of the previous gains.
This makes the next proposal consistent with the proposal listing.

Changed the calling syntax for find_next_proposal() to make it
consistent with the other model functions. Old syntax was:
prop.id = Proposal.find_next_proposal(prop.id, prop.type.id, person.id)
New syntax uses objects in and out rather than ids:
prop = Proposal.find_next_proposal(prop, person)