Fix #306 - review comment columns too small
The issue is a little more complex than reported.
The table listing the review comments was a variable width table that
took the full width of the active body area. Normally that meant that
the comments took roughly a third of the width, private comments another
third. However if one column had a long string in it, like a link, then
it stole column width from the others leading to a very narrow column as
depicted in the bug report.

However, even at its optimum the arrangement is poor. The area to enter
comments is substantial encouraging a paragraph or two. We want to
encourage this level of detail. Having a third of the active area given
to the comments is poor and rather hard to read.

This commit fixes all this by abusing the standard table model.
Each review gets three rows of the table. The first row has the
reviewer, score and (optionally) the stream. The second row has the
comments, the third the private comments.

The table is abused further by making the comments use up all the table
columns except the first. So typically the comments section takes over
five columns, 90% of the active area width.

In practice comments are now nice solid paragraphs and the world is a
better place.