Fix #76 - Detect duplicate submissions
This commit detects proposals with similar titles and abstract texts.
The test is performed when a reviewer loads the proposal, if any
potentially duplicate proposals are found a list is provided to them.

This involved putting a new function in the proposal model, adding the
functionality to the controller and template. Both model and functional
testing are provided.

The comparisons are performed using postgresql functions which are
provided with the program but typically packaged as extras. The alembic
version enables these function groups.

Debian based systems will need to install postgresql-contrib-9.4 or
similar. Other distributions may also require the contrib or extra
package.

A fallback is also provided for non-postgresql systems using identical
matches.

Performance, benchmarking and optimisation was performed against the
2016 submission set.