Fixed #33 - sort by type on /proposal/approve
Issue requested sorting by type on the basis that updates were typically
applied to one type at a time.

Initial fix was to just apply a sort.

Then transitioned to providing a heading and table per type, this is
clearer to use, more consistent with other pages, and allowed the easy
creation of a contents menu for faster navigation.