Fix #166 - Remove miniconf links after approval
The issue is that after miniconfs are decided they are still listed on
the to-be-reviewed page.

I updated the choose review page to only list proposals that had a
status of pending review. Also updated find_next_proposal to only choose
pending review status too.

To wrap things up any proposal type without an eligible pending proposal
is hidden.

To prevent this being confusing the review summary page also displays
the status as a score if it isn't pending review.